+++ CHANGELOG.md
@@ -12,6 +12,7 @@ New methods:
 - Added ``offset_curve`` method from shapely to GeoSeries/GeoDataframe (#2902).
 - Added ``shortest_line`` method from shapely to GeoSeries/GeoDataframe (#2960).
 - Added ``minimum_rotated_rectangle`` method from shapely to GeoSeries/GeoDataframe (#2541).
+- Added ``count_coordinates`` method from shapely to GeoSeries/GeoDataframe (#3026).
 
 New features and improvements: