+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Notes on dependencies:
   currently supported is shapely >= 2. As a consequence, spatial indexing based on the
   rtree package has also been removed. (#3035)
 
+New methods:
+
+- Added ``minimum_clearance`` method from shapely to GeoSeries/GeoDataframe (#2989).
 
 Bug fixes:
 - Fix `GeoDataFrame.merge()` incorrectly returning a `DataFrame` instead of a